Install figlet from source
The vms don't have very many pacakges available to them. Figlet, which
I want to use to make text banners is one of them. So I've added a function
to wget it from the figlet ftp server and build it from source. Please note
that this is a huge waste of time.